Fix certificate has expired (replace node-fetch with electron-fetch)
