+++ package.json
@@ -53,6 +53,7 @@
     "electron": "13.5.0",
     "electron-builder": "22.11.11",
     "electron-context-menu": "3.1.1",
+    "electron-fetch": "1.7.4",
     "electron-notarize": "1.1.1",
     "electron-packager": "15.4.0",
     "electron-settings": "4.0.2",
